--- application-bemyplan/src/main/java/com/deploy/bemyplan/controller/plan/dto/request/RetrievePlansRequest.java
@@ -1,19 +1,0 @@
-package com.deploy.bemyplan.controller.plan.dto.request;
-
-import com.deploy.bemyplan.domain.plan.RegionCategory;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import org.jetbrains.annotations.Nullable;
-
-@ToString
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class RetrievePlansRequest {
-    @Nullable
-    private RegionCategory region;
-
-}
