--- application-bemyplan/src/main/java/com/deploy/bemyplan/service/plan/dto/request/RetrieveMyBookmarkListRequestDto.java
@@ -1,21 +1,0 @@
-package com.deploy.bemyplan.service.plan.dto.request;
-
-import lombok.*;
-import org.jetbrains.annotations.Nullable;
-
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-
-@ToString
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class RetrieveMyBookmarkListRequestDto {
-
-    @Min(value = 1, message = "{common.size.min}")
-    @Max(value = 100, message = "{common.size.max}")
-    private int size;
-
-    @Nullable
-    private Long lastScrapId;
-}
