--- application-bemyplan/src/test/java/com/deploy/bemyplan/controller/plan/PlanControllerTest.java
@@ -1,82 +1,0 @@
-package com.deploy.bemyplan.controller.plan;
-
-import com.deploy.bemyplan.domain.plan.Plan;
-import com.deploy.bemyplan.domain.plan.PlanStatus;
-import com.deploy.bemyplan.domain.plan.RcmndStatus;
-import com.deploy.bemyplan.domain.plan.Region;
-import com.deploy.bemyplan.domain.plan.RegionCategory;
-import com.deploy.bemyplan.domain.plan.TagInfo;
-import com.deploy.bemyplan.service.plan.PlanService;
-import com.deploy.bemyplan.service.plan.dto.response.PlanPreviewResponseDto;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-
-import java.util.Collections;
-import java.util.List;
-
-import static com.deploy.bemyplan.domain.plan.Plan.newInstance;
-import static org.hamcrest.Matchers.equalTo;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-class PlanControllerTest {
-    private MockMvc mockMvc;
-
-    @Mock
-    private PlanService stubPlanService;
-
-    @BeforeEach
-    void setUp() {
-        MockitoAnnotations.openMocks(this);
-
-        mockMvc = MockMvcBuilders.standaloneSetup(new PlanController(stubPlanService))
-                .build();
-    }
-
-    @Test
-    void getPlanPreviewReturnsOkHttpStatus() throws Exception {
-        mockMvc.perform(get("/api/v2/plans/1/preview"))
-                .andExpect(status().isOk());
-    }
-
-    @Test
-    void getPlanPreviewPassesPlanIdToService() throws Exception {
-        mockMvc.perform(get("/api/v2/plans/1/preview"));
-
-        verify(stubPlanService, times(1)).getPlanPreview(1L);
-    }
-
-    @Test
-    void getPlanPreviewReturnsResponse() throws Exception {
-        Plan givenPlan = newInstance(1L, RegionCategory.JEJU, Region.JEJUALL, "thumbnail", "title", "description", TagInfo.testBuilder().build(), 2000, PlanStatus.ACTIVE, RcmndStatus.NONE, Collections.emptyList(), Collections.emptyList());
-        PlanPreviewResponseDto response = PlanPreviewResponseDto.of(givenPlan, List.of("image.png", "image2.png"));
-        given(stubPlanService.getPlanPreview(any())).willReturn(response);
-
-        mockMvc.perform(get("/api/v2/plans/{planId}/preview", 1L))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.planId", equalTo(response.getPlanId())))
-                .andExpect(jsonPath("$.title", equalTo(response.getTitle())))
-                .andExpect(jsonPath("$.description", equalTo(response.getDescription())))
-                .andExpect(jsonPath("$.price", equalTo(response.getPrice())))
-                .andExpect(jsonPath("$.month", equalTo(response.getMonth())))
-                .andExpect(jsonPath("$.totalDay", equalTo(response.getTotalDay())))
-                .andExpect(jsonPath("$.theme", equalTo(response.getTheme())))
-                .andExpect(jsonPath("$.budget", equalTo(response.getBudget())))
-                .andExpect(jsonPath("$.hashtag", equalTo(response.getHashtag())))
-                .andExpect(jsonPath("$.recommendTarget", equalTo(response.getRecommendTarget())))
-                .andExpect(jsonPath("$.travelMobility", equalTo(response.getTravelMobility())))
-                .andExpect(jsonPath("$.restaurantCount", equalTo(response.getRestaurantCount())))
-                .andExpect(jsonPath("$.spotCount", equalTo(response.getSpotCount())))
-                .andExpect(jsonPath("$.thumbnail", equalTo(response.getThumbnail())))
-                .andExpect(jsonPath("$.travelPartner", equalTo(response.getTravelPartner())));
-    }
-}
