+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/plan/service/dto/request/RetrieveMyBookmarkListRequestDto.java
@@ -1,0 +1,25 @@
+package com.deploy.bemyplan.plan.service.dto.request;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import org.jetbrains.annotations.Nullable;
+
+import javax.validation.constraints.Max;
+import javax.validation.constraints.Min;
+
+@ToString
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class RetrieveMyBookmarkListRequestDto {
+
+    @Min(value = 1, message = "{common.size.min}")
+    @Max(value = 100, message = "{common.size.max}")
+    private int size;
+
+    @Nullable
+    private Long lastScrapId;
+}
